changed Weiqi.Web for Weiqi.Desktop
