+++ Weiqi.Desktop/App.xaml
