+++ Weiqi.Desktop/AssemblyInfo.cs
