+++ Weiqi.Desktop/Interfaces/IStone.cs
@@ -1,0 +1,6 @@
+namespace Weiqi.Desktop.Interfaces;
+
+public interface IStone
+{
+    void Draw(double x, double y, double size);
+}
