--- Weiqi.Web/Interfaces/IStone.cs
@@ -1,6 +1,0 @@
-namespace Weiqi.Web.Interfaces;
-
-public interface IStone
-{
-    void Draw(double x, double y, double size);
-}
