--- Weiqi.Web/Models/CanvasStone.cs
@@ -1,28 +1,0 @@
-using System.Windows.Controls;
-using System.Windows.Media;
-using System.Windows.Shapes;
-using Weiqi.Engine.Models;
-
-namespace Weiqi.Web.Models
-{
-    public class CanvasStone : StoneRepresentation
-    {
-        public CanvasStone(Stone stone) : base(stone) { }
-
-        public override void Draw(Canvas canvas, double x, double y, double size)
-        {
-            Ellipse stoneEllipse = new Ellipse
-            {
-                Width = size - 4,
-                Height = size - 4,
-                Fill = stone == Stone.Black ? Brushes.Black : Brushes.White,
-                HorizontalAlignment = System.Windows.HorizontalAlignment.Center,
-                VerticalAlignment = System.Windows.VerticalAlignment.Center
-            };
-
-            Canvas.SetLeft(stoneEllipse, x - stoneEllipse.Width / 2);
-            Canvas.SetTop(stoneEllipse, y - stoneEllipse.Height / 2);
-            canvas.Children.Add(stoneEllipse);
-        }
-    }
-}
